+++ tests/python_package_test/test_engine.py
@@ -18,8 +18,8 @@ from sklearn.model_selection import GroupKFold, TimeSeriesSplit, train_test_spli
 
 import lightgbm as lgb
 
-from .utils import (load_boston, load_breast_cancer, load_digits, load_iris, make_synthetic_regression, pickle_obj,
-                    sklearn_multiclass_custom_objective, softmax, unpickle_obj)
+from .utils import (load_boston, load_breast_cancer, load_digits, load_iris, make_synthetic_regression,
+                    sklearn_multiclass_custom_objective, softmax)
 
 decreasing_generator = itertools.count(0, -1)
 
@@ -3424,19 +3424,3 @@ def test_pandas_nullable_dtypes():
 
     # test equal predictions
     np.testing.assert_allclose(preds, preds_nullable_dtypes)
-
-
-@pytest.mark.parametrize('serializer', ["pickle", "joblib", "cloudpickle"])
-def test_early_stopping_callback_is_picklable(serializer, tmp_path):
-    callback = lgb.early_stopping(stopping_rounds=5)
-    tmp_file = tmp_path / "early_stopping.pkl"
-    pickle_obj(
-        obj=callback,
-        filepath=tmp_file,
-        serializer=serializer
-    )
-    callback_from_disk = unpickle_obj(
-        filepath=tmp_file,
-        serializer=serializer
-    )
-    assert callback.stopping_rounds == callback_from_disk.stopping_rounds
